+++ my_server/lib/my_server.dart
@@ -5,7 +5,6 @@ library my_server;
 
 export 'dart:async';
 export 'dart:io';
-
+export 'package:sqljocky5/sqljocky.dart';
 export 'package:aqueduct/aqueduct.dart';
-
 export 'channel.dart';